@@ -30,52 +30,42 @@ abstract class NetworkBoundResource<LocalType, RemoteType> : FlowableOnSubscribe
 
     fun fetch(isForced: Boolean) {
         getLocal()
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .flatMap {
-                    emitter?.onNext(success(it))
-                    return@flatMap getRemote()
-                }.map(mapper())
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .flatMap {
-                    saveCallResult(it, isForced)
-                    return@flatMap getLocal()
-                }
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .subscribe({
-                    emitter?.onNext(success(it))
-                }, {
-                    emitter?.onNext(Resource.error(it))
-                })
-
-//        getRemote().map(mapper())
-//                .subscribeOn(Schedulers.io())
-//                .observeOn(Schedulers.io())
-//                .subscribe({ localTypeData ->
-//                    saveCallResult(localTypeData, isForced)
-//                }, {
-//                    emitter?.onNext(Resource.error(it))
-//                })
-
+            .subscribeOn(Schedulers.io())
+            .observeOn(Schedulers.io())
+            .flatMap {
+                emitter?.onNext(success(it))
+                return@flatMap getRemote()
+            }.map(mapper())
+            .subscribeOn(Schedulers.io())
+            .observeOn(Schedulers.io())
+            .flatMap {
+                saveCallResult(it, isForced)
+                return@flatMap getLocal()
+            }
+            .subscribeOn(Schedulers.io())
+            .observeOn(Schedulers.io())
+            .subscribe({
+                emitter?.onNext(success(it))
+            }, {
+                emitter?.onNext(Resource.error(it))
+            })
     }
 
     fun getRemoteData(isForced: Boolean) {
         emitter?.onNext(loading())
         getRemote().map(mapper())
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .flatMap {
-                    saveCallResult(it, isForced)
-                    return@flatMap getLocal()
-                }
-                .subscribeOn(Schedulers.io())
-                .observeOn(Schedulers.io())
-                .subscribe({
-                    emitter?.onNext(success(it))
-                }, {
-                    emitter?.onNext(Resource.error(it))
-                })
+            .subscribeOn(Schedulers.io())
+            .observeOn(Schedulers.io())
+            .flatMap {
+                saveCallResult(it, isForced)
+                return@flatMap getLocal()
+            }
+            .subscribeOn(Schedulers.io())
+            .observeOn(Schedulers.io())
+            .subscribe({
+                emitter?.onNext(success(it))
+            }, {
+                emitter?.onNext(Resource.error(it))
+            })
     }
 }
