@@ -1,7 +1,7 @@
 package com.chintansoni.android.repositorypattern.model.local
 
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.RoomDatabase
 import com.chintansoni.android.repositorypattern.model.local.dao.UserDao
 import com.chintansoni.android.repositorypattern.model.local.entity.User
 
