@@ -1,9 +1,9 @@
 package com.chintansoni.android.repositorypattern.model.local.entity
 
-import android.arch.persistence.room.Embedded
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
 import android.os.Parcelable
+import androidx.room.Embedded
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 import com.chintansoni.android.repositorypattern.model.local.DatabaseConstants
 import com.chintansoni.android.repositorypattern.model.remote.response.Dob
 import com.chintansoni.android.repositorypattern.model.remote.response.Location
