@@ -1,7 +1,7 @@
 package com.chintansoni.android.repositorypattern.util
 
-import android.databinding.BindingAdapter
-import android.widget.ImageView
+import androidx.appcompat.widget.AppCompatImageView
+import androidx.databinding.BindingAdapter
 import com.bumptech.glide.load.engine.DiskCacheStrategy
 import com.chintansoni.android.repositorypattern.R
 import com.chintansoni.android.repositorypattern.model.remote.response.Location
@@ -10,15 +10,17 @@ import com.chintansoni.android.repositorypattern.model.remote.response.Name
 object BindingUtils {
     @JvmStatic
     @BindingAdapter("app:imageUrl", "app:gender")
-    fun loadImage(view: ImageView, url: String, gender: String) {
+    fun loadImage(view: AppCompatImageView, url: String, gender: String) {
         GlideApp.with(view.context)
-                .load(url)
-                .placeholder(when (gender) {
+            .load(url)
+            .placeholder(
+                when (gender) {
                     "male" -> R.drawable.ic_male_placeholder
                     else -> R.drawable.ic_female_placeholder
-                })
-                .diskCacheStrategy(DiskCacheStrategy.AUTOMATIC)
-                .into(view)
+                }
+            )
+            .diskCacheStrategy(DiskCacheStrategy.AUTOMATIC)
+            .into(view)
     }
 
     @JvmStatic
@@ -34,10 +36,5 @@ object BindingUtils {
     @JvmStatic
     fun getDate(s: String): String? {
         return s.split("T")[0]
-    }
-
-    @JvmStatic
-    fun loadAvatar() {
-
     }
 }
