@@ -1,23 +1,25 @@
 package com.chintansoni.android.repositorypattern.view.activity
 
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.chintansoni.android.repositorypattern.R
 import com.chintansoni.android.repositorypattern.model.local.entity.User
 import com.chintansoni.android.repositorypattern.view.adapter.UserRecyclerAdapter
 import com.chintansoni.android.repositorypattern.view.fragment.DetailFragment
 import com.chintansoni.android.repositorypattern.view.fragment.ListFragment
-import dagger.android.support.DaggerAppCompatActivity
 
-class UsersActivity : DaggerAppCompatActivity(), UserRecyclerAdapter.ItemTouchListener {
+class UsersActivity : AppCompatActivity(), UserRecyclerAdapter.ItemTouchListener {
     override fun onItemClick(user: User) {
         supportFragmentManager.beginTransaction()
-                .setCustomAnimations(R.anim.slide_up,
-                        R.anim.slide_down,
-                        R.anim.slide_up,
-                        R.anim.slide_down)
-                .add(R.id.container, DetailFragment.newInstance(user))
-                .addToBackStack(detailFragmentTag)
-                .commit()
+            .setCustomAnimations(
+                R.anim.slide_up,
+                R.anim.slide_down,
+                R.anim.slide_up,
+                R.anim.slide_down
+            )
+            .add(R.id.container, DetailFragment.newInstance(user))
+            .addToBackStack(detailFragmentTag)
+            .commit()
     }
 
     companion object {
@@ -29,8 +31,8 @@ class UsersActivity : DaggerAppCompatActivity(), UserRecyclerAdapter.ItemTouchLi
         setContentView(R.layout.users_activity)
         if (savedInstanceState == null) {
             supportFragmentManager.beginTransaction()
-                    .replace(R.id.container, ListFragment.newInstance())
-                    .commit()
+                .replace(R.id.container, ListFragment.newInstance())
+                .commit()
         }
     }
 }
