@@ -1,6 +1,6 @@
 package com.chintansoni.android.repositorypattern.view.viewholder
 
-import android.support.v7.widget.RecyclerView
 import com.chintansoni.android.repositorypattern.databinding.ItemLoadingBinding
 
-class LoaderViewHolder(mLoadingBinding: ItemLoadingBinding) : RecyclerView.ViewHolder(mLoadingBinding.root)
+class LoaderViewHolder(mLoadingBinding: ItemLoadingBinding) :
+    androidx.recyclerview.widget.RecyclerView.ViewHolder(mLoadingBinding.root)
